Conflicts:
	MC02-Group9/.DS_Store

+++ MC02-Group9/MC02-Group9/Manager/CoreDataManager.swift
@@ -44,7 +44,35 @@ class CoreDataManager{
         
     }
     
+    func updateLog(log: Log, logFire:LogFire){
+        log.bg_check_result = logFire.bg_check_result
+        log.action = logFire.action
+        
+        do{
+            try self.context.save()
+        }
+        catch {
+            
+        }
+        
+    }
+    
     func batalkan(logToRemove: Log){
+        if logToRemove.type == 1 {
+            // BG Logic
+//            logToRemove.bg_check_result = "-1"
+            do{
+                try self.context.save()
+            }
+            catch {
+                
+            }
+            
+            return
+        }
+        
+        MigrateFirestoreToCoreData.migrateFirestoreToCoreData.removeLogToFirestore(id: logToRemove.log_id!)
+        
         // remove log
         self.context.delete(logToRemove)
         
@@ -53,6 +81,9 @@ class CoreDataManager{
         }catch{
             
         }
+        
+        
+        
     }
     
 //    func saveUser(user_role:Int16){
@@ -68,11 +99,57 @@ class CoreDataManager{
 //        }
 //    }
     
+    func getLogRealIdx(med: Medicine_Time) -> Int {
+        
+        guard let getLogs = logs else{
+            return -1
+        }
+        
+        var idx = 0
+
+        for l in getLogs{ //menentukan log bg atau med
+            if l.type == 0 && med.time == l.time && med.medicine!.name == l.medicine_name {
+                return idx
+            }
+            
+            idx += 1
+        }
+        
+        return -1
+    }
+    
+    func getLogRealIdx(log: Log) -> Int {
+        
+        guard let getLogs = logs else{
+            return -1
+        }
+        
+        var idx = 0
+
+        for l in getLogs{ //menentukan log bg atau med
+            if log.type == l.type && log.time == l.time{
+                return idx
+            }
+            
+            idx += 1
+        }
+        
+        return -1
+    }
+    
     func makeLogInit() -> Log {
         
         let newLog = Log(context: self.context)
+        newLog.action = ""
+        newLog.bg_check_result = "-1"
+        newLog.date = Date()
+        newLog.dateTake = Date()
         newLog.log_id = UUID().uuidString
+        newLog.medicine_name = ""
+        newLog.time = ""
         newLog.type = 0 //med
+        
+        
         
         return newLog
         
@@ -115,46 +192,32 @@ class CoreDataManager{
         }catch{
             
         }
+        
+        //Firestore
+        MigrateFirestoreToCoreData.migrateFirestoreToCoreData.addNewLogToFirestore(log: log)
     }
     
-    func simpanBG(daySelected: Date,bGResult:String,bg:BG){
-        //change daySelected to String
-        let formatter = DateFormatter()
-        formatter.locale = Locale(identifier: "en_gb")
-        formatter.dateFormat = "dd MMM yyyy"
-        let tanggal = formatter.string(from: daySelected)
-        // print(tanggal)
+    func simpanBG(date: Date, time: String ,bGResult:String){
+        // New
         
-        // Create String
-        let time = bg.bg_time!
-        let hour = time[..<time.index(time.startIndex, offsetBy: 2)]
-        let minutes = time[time.index(time.startIndex, offsetBy: 3)...]
-        let string = ("\(tanggal) \(hour):\(minutes):00 +0700")
-        print(string)
-        // 29 October 2019 20:15:55 +0200
-
+        let newLog = makeLogInit()
+        newLog.date = date
+        newLog.time = time
+        newLog.bg_check_result = bGResult
+        newLog.type = 1
+        newLog.action = "Take"
         
-        // Create Date Formatter
-        let dateFormatter = DateFormatter()
-
-        // Set Date Format
-        dateFormatter.dateFormat = "dd MMM yyyy HH:mm:ss Z"
-        // Convert String to Date
-        print("\(dateFormatter.date(from: string)!) ubah ke UTC")
         
-        let log = makeLogInit()
-        log.date = dateFormatter.date(from: string)
-        log.dateTake = dateFormatter.date(from: string)
-        log.action = "Take"
-        log.bg_check_result = bGResult
-        log.type = 1
-        print("INI BG RESULTNYA \(log.bg_check_result)")
+        print("log baru \(newLog.bg_check_result!)")
         
         do{
             try self.context.save()
         }catch{
             
         }
+        
+        //Firestore
+        MigrateFirestoreToCoreData.migrateFirestoreToCoreData.addNewLogToFirestore(log: newLog)
     }
     
     func lewati(daySelected: Date, indexPath: IndexPath){
@@ -194,6 +257,9 @@ class CoreDataManager{
         }catch{
             
         }
+        
+        //Firestore
+        MigrateFirestoreToCoreData.migrateFirestoreToCoreData.addNewLogToFirestore(log: log)
     }
     
     func pilihWaktu(daySelected: Date, indexPath: IndexPath, myDatePicker: UIDatePicker){
@@ -240,6 +306,9 @@ class CoreDataManager{
         }catch{
             
         }
+        
+        //Firestore
+        MigrateFirestoreToCoreData.migrateFirestoreToCoreData.addNewLogToFirestore(log: log)
     }
     
     func bgLog(bgDate:Date,bgTime:String){
@@ -255,6 +324,9 @@ class CoreDataManager{
         }catch{
             
         }
+        
+        //Firestore
+        MigrateFirestoreToCoreData.migrateFirestoreToCoreData.addNewLogToFirestore(log: log)
     }
     
     func checkBG(daySelected: Date, indexPath: IndexPath){
@@ -323,6 +395,9 @@ class CoreDataManager{
         }catch{
             
         }
+        
+        //Firestore
+        MigrateFirestoreToCoreData.migrateFirestoreToCoreData.addNewLogToFirestore(log: log)
     }
     
     func fetchBG(){
@@ -401,6 +476,21 @@ class CoreDataManager{
         }catch{
             
         }
+    }
+    
+    func fetchAllLogs() -> [Log]{
+        
+        do {
+            let request = Log.fetchRequest() as NSFetchRequest<Log>
+            let logs = try context.fetch(request)
+            
+            return logs
+        }catch{
+            
+            
+        }
+        
+        return [Log]()
     }
     
     func fetchLogs(tableView: UITableView, daySelected: Date){
